♻️ refactor: move PLUGIN_CATEGORIES data into config.yaml
- Add plugin_categories section to config.yaml
- Load and parse into dict[str, CategoryInfo] in config.py
- Delete plugin_categories.py, update all imports to use app.config.config
